# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian Advisory API",
    "description": "Consultation booking, payments, client and consultant portals, commerce and content APIs.\n\nAll responses use a single envelope: `{ success: true, data }` or `{ success: false, error: { code, message } }`.\n\nAuthenticate with `Authorization: Bearer <accessToken>`. Access tokens are short-lived; refresh them at `POST /api/auth/refresh`, which reads an httpOnly cookie.",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@meridianadvisory.co.ke"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/branding/{asset}": {
      "get": {
        "summary": "The current logo, dark logo or favicon",
        "tags": [
          "System"
        ],
        "description": "Public and stable: the URL does not change when the image behind it does. Cached for a year against its version parameter, which changes on every upload.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "asset",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/branding": {
      "get": {
        "summary": "Current brand assets",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/admin/branding/{asset}": {
      "post": {
        "summary": "Replace a brand image",
        "tags": [
          "Admin"
        ],
        "description": "Multipart upload of a PNG, JPEG, WebP or SVG, 2MB or smaller.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "asset",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Remove a brand image and fall back to the built-in mark",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "asset",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "development"
    }
  ],
...
}